brute forceing my way past dumb computer brain
